+++ src/ionic-serve.ts
@@ -1,7 +1,6 @@
 import * as vscode from 'vscode';
 
 import { InternalCommand } from './command-name';
-import { ionicState } from './ionic-tree-provider';
 import { MonoRepoType } from './monorepo';
 import { preflightNPMCheck } from './node-commands';
 import { Project } from './project';